+++ src/module.ts
@@ -3,12 +3,13 @@ import {
     createResolver,
     defineNuxtModule,
     addTypeTemplate,
-    addComponent
+    addComponentsDir
 } from '@nuxt/kit'
-import {join} from 'node:path'
 
-import {buildIcons} from './runtime/build'
-import {generateIconsTypes} from './types/generate-icons-types'
+import { join } from 'node:path'
+
+import { buildIcons } from './runtime/build'
+import { generateIconsTypes } from './types/generate-icons-types'
 
 const PACKAGE_NAME = 'nuxt-feather-icons'
 
@@ -20,17 +21,12 @@ export interface ModuleOptions {
     prefix?: string
 }
 
-interface ModuleIconsNames {
-    componentName: string
-    componentPascalName: string
-}
-
 export default defineNuxtModule<ModuleOptions>({
     meta: {
         name: PACKAGE_NAME,
         configKey: 'nuxtFeatherIcons',
         compatibility: {
-            nuxt: '>=3.0.0',
+            nuxt: '>=3.0.0'
         }
     },
 
@@ -40,21 +36,20 @@ export default defineNuxtModule<ModuleOptions>({
 
     async setup(options, nuxt) {
         const logger = useLogger(PACKAGE_NAME)
+        const { resolve } = createResolver(import.meta.url)
 
         logger.info('Generating Feather icons components...')
 
         const icons = await buildIcons(nuxt)
 
-        for (const icon of icons) {
-            const componentName = `${options.prefix}${icon.componentPascalName}`
+        const componentsDir = join(nuxt.options.buildDir, 'feather-icons')
 
-            addComponent({
-                name: componentName,
-                export: 'default',
-                filePath: `#build/feather-icons/${icon.componentPascalName}`,
-                chunkName: `feather-${componentName}`,
-            })
-        }
+        addComponentsDir({
+            path: componentsDir,
+            pathPrefix: false,
+            extensions: ['js'],
+            prefix: options.prefix
+        })
 
         addTypeTemplate({
             filename: 'types/nuxt-feather-icons.d.ts',
@@ -65,8 +60,8 @@ export default defineNuxtModule<ModuleOptions>({
             prefix: options.prefix || ''
         }
 
-        nuxt.hook('prepare:types', ({references}) => {
-            references.push({path: 'types/nuxt-feather-icons.d.ts'})
+        nuxt.hook('prepare:types', ({ references }) => {
+            references.push({ path: 'types/nuxt-feather-icons.d.ts' })
         })
 
         logger.success(`${icons.length} Feather icons registered`)